`object_player`->`model_player`:add_head
