+++ include/mesh/mesh_player_head.h
@@ -1,0 +1,10 @@
+#ifndef __mesh_player_head_h
+#define __mesh_player_head_h
+
+#include <metil_mesh/metil_mesh.h>
+
+void mesh_player_head_initialize(
+  struct metil_mesh*
+);
+
+#endif
