+++ sources/mesh/mesh_player.c
@@ -1,6 +1,7 @@
 #include <mesh/mesh_player.h>
 
 #include <math_c_vector.h>
+
 #include <metil_mesh/metil_mesh.h>
 
 #include <math.h>
